+++ core/src/com/neves6/piazzapanic/gamemaster/GameMaster.java
@@ -1,4 +1,4 @@
-package com.neves6.piazzapanic;
+package com.neves6.piazzapanic.gamemaster;
 
 /** GameMaster class. */
 public class GameMaster {
