+++ core/src/com/neves6/piazzapanic/gamemechanisms/Utility.java
@@ -1,4 +1,4 @@
-package com.neves6.piazzapanic;
+package com.neves6.piazzapanic.gamemechanisms;
 
 import java.io.BufferedWriter;
 import java.io.File;
