+++ core/src/com/neves6/piazzapanic/people/Chef.java
@@ -1,9 +1,12 @@
-package com.neves6.piazzapanic;
+package com.neves6.piazzapanic.people;
 
-import com.badlogic.gdx.graphics.Texture;
+import com.badlogic.gdx.graphics.Texture;import com.neves6.piazzapanic.gamemechanisms.Machine;
 import java.util.Stack;
 
-/** Chef subclass. */
+/**
+ * Chef class which helps the user cook items and it
+ * inherits methods from the person class.
+ */
 public class Chef extends Person {
   private boolean isStickied;
   private Stack<String> inventory;
@@ -54,47 +57,82 @@ public class Chef extends Person {
     this.txNow = txDown;
   }
 
+  /**
+   * Getter method used to see whether chef can move or not.
+   * @return Can the chef move.
+   */
   public boolean getIsStickied() {
     return isStickied;
   }
 
+  /**
+   * Getter method used to set which texture is being used
+   * to render the chef.
+   * @return Texture being used to render the chef.
+   */
   public Texture getTxNow() {
     return txNow;
   }
 
+  /**
+   * Setter method to set whether the chef is stuck to
+   * one coordinates.
+   * @param flag Whether the chef is stuck to one coordinates.
+   */
   public void setIsStickied(boolean flag) {
     this.isStickied = flag;
   }
 
+  /**
+   * Setter method to show what the chef is currently using.
+   * @param machine The machine that the chef is interacting with.
+   */
   public void setMachineInteractingWith(Machine machine) {
     this.machineInteractingWith = machine;
   }
 
+  /**
+   * Getter method to see what machine the chef is using.
+   * @return The machine that the chef is interacting with.
+   */
   public Machine getMachineInteractingWith() {
     return machineInteractingWith;
   }
 
-  // Inventory management
+
+  /***
+   * Getter method to see what items the chef has.
+   * @return The items that the chef has in its possession.
+   */
   public Stack<String> getInventory() {
     return inventory;
   }
 
+  /**
+   * Add a new item to the top of the chef inventory.
+   * @param item A ingredient to be added to the inventory.
+   */
   public void addToInventory(String item) {
     this.inventory.push(item);
   }
 
+  /**
+   * Removes the item at the top of the chefs inventory.
+   */
   public void removeTopFromInventory() {
     this.inventory.pop();
   }
 
-  // Facing
+  /**
+   * Getter method for which way the chef is facing.
+   * @return Which way the chef is facing.
+   */
   public String getFacing() {
     return facing;
   }
 
   /**
    * Sets the facing direction of the chef, then changes the current texture to match.
-   *
    * @param facing sprite facing direction.
    */
   public void setFacing(String facing) {
