+++ core/src/com/neves6/piazzapanic/screens/CreditsScreen.java
@@ -1,4 +1,4 @@
-package com.neves6.piazzapanic;
+package com.neves6.piazzapanic.screens;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.InputProcessor;
@@ -8,6 +8,9 @@ import com.badlogic.gdx.graphics.OrthographicCamera;
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.BitmapFont;
 
+/**
+ * Display once the game has been completed by the user.
+ */
 public class CreditsScreen extends ScreenAdapter implements InputProcessor {
   PiazzaPanicGame game;
   OrthographicCamera camera;
@@ -16,12 +19,20 @@ public class CreditsScreen extends ScreenAdapter implements InputProcessor {
   int winWidth;
   int winHeight;
 
+  /**
+   * Constructor method which sets up the fonts and textures to display the
+   * credits.
+   * @param game
+   */
   public CreditsScreen(PiazzaPanicGame game) {
     this.game = game;
     font = new BitmapFont(Gdx.files.internal("fonts/IBM_Plex_Mono_SemiBold.fnt"));
     credits = new Texture(Gdx.files.internal("credits.png"));
   }
 
+  /**
+   * Inherited show method which sets up screen dimensions.
+   */
   @Override
   public void show() {
     Gdx.input.setInputProcessor(this);
@@ -30,6 +41,10 @@ public class CreditsScreen extends ScreenAdapter implements InputProcessor {
     camera.setToOrtho(false, Gdx.graphics.getWidth(), Gdx.graphics.getHeight());
   }
 
+  /**
+   * Draws the credits image onto the screen.
+   * @param delta The time in seconds since the last render.
+   */
   @Override
   public void render(float delta) {
     Gdx.gl20.glViewport(0, 0, Gdx.graphics.getWidth(), Gdx.graphics.getHeight());
@@ -46,12 +61,20 @@ public class CreditsScreen extends ScreenAdapter implements InputProcessor {
     game.getBatch().end();
   }
 
+  /**
+   * A function which can resize the window if it changes.
+   * @param width Horizontal size of the screen.
+   * @param height Vertical size of the screen.
+   */
   @Override
   public void resize(int width, int height) {
     super.resize(width, height);
     camera.setToOrtho(false, width, height);
   }
 
+  /**
+   * What to do when the screen is exited.
+   */
   @Override
   public void hide() {
     super.dispose();
@@ -61,8 +84,10 @@ public class CreditsScreen extends ScreenAdapter implements InputProcessor {
   }
 
   /**
-   * @param keycode one of the constants in {@link Input.Keys}
-   * @return
+   * When the user presses a single key down.
+   * @param keycode a one to one mapping between the user keyboard
+   *                and numbers.
+   * @return true
    */
   @Override
   public boolean keyDown(int keycode) {
@@ -71,8 +96,9 @@ public class CreditsScreen extends ScreenAdapter implements InputProcessor {
   }
 
   /**
+   * UNUSED (NEED TO INHERIT CLASS)
    * @param keycode one of the constants in {@link Input.Keys}
-   * @return
+   * @return false
    */
   @Override
   public boolean keyUp(int keycode) {
@@ -80,8 +106,9 @@ public class CreditsScreen extends ScreenAdapter implements InputProcessor {
   }
 
   /**
+   *
    * @param character The character
-   * @return
+   * @return false
    */
   @Override
   public boolean keyTyped(char character) {
