+++ core/src/com/neves6/piazzapanic/screens/GameWinScreen.java
@@ -1,4 +1,4 @@
-package com.neves6.piazzapanic;
+package com.neves6.piazzapanic.screens;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.ScreenAdapter;
@@ -14,29 +14,28 @@ import com.badlogic.gdx.scenes.scene2d.ui.Skin;
 import com.badlogic.gdx.scenes.scene2d.ui.TextButton;
 import com.badlogic.gdx.scenes.scene2d.utils.ChangeListener;
 
-public class LevelSelectorScreen extends ScreenAdapter {
+public class GameWinScreen extends ScreenAdapter {
   PiazzaPanicGame game;
   OrthographicCamera camera;
   SpriteBatch batch;
   BitmapFont font;
   Texture bg;
-  Texture lock;
   int winWidth;
   int winHeight;
   float bgScaleFactor;
   Stage stage;
-  TextButton level1Button;
-  TextButton level2Button;
-  TextButton level3Button;
+  TextButton creditsButton;
+  TextButton titleButton;
   TextButton.TextButtonStyle buttonStyle;
   Skin skin;
   TextureAtlas atlas;
+  int completionTime;
 
-  public LevelSelectorScreen(PiazzaPanicGame game) {
+  public GameWinScreen(PiazzaPanicGame game, int completionTime) {
     this.game = game;
+    this.completionTime = completionTime;
     font = new BitmapFont(Gdx.files.internal("fonts/IBM_Plex_Mono_SemiBold.fnt"));
     bg = new Texture(Gdx.files.internal("title_screen_large-min.png"));
-    lock = new Texture(Gdx.files.internal("levellocked.png"));
   }
 
   @Override
@@ -48,50 +47,39 @@ public class LevelSelectorScreen extends ScreenAdapter {
     stage = new Stage();
     Gdx.input.setInputProcessor(stage);
     skin = new Skin();
-    atlas = new TextureAtlas(Gdx.files.internal("buttons/levelselector/levelselector.atlas"));
+    atlas = new TextureAtlas(Gdx.files.internal("buttons/title/unnamed.atlas"));
     skin.addRegions(atlas);
     buttonStyle = new TextButton.TextButtonStyle();
     buttonStyle.font = font;
-    buttonStyle.up = skin.getDrawable("black_alpha_square");
-    buttonStyle.down = skin.getDrawable("black_alpha_square");
-    buttonStyle.checked = skin.getDrawable("black_alpha_square");
-    level1Button = new TextButton("Level 1", buttonStyle);
-    level1Button.setPosition(
-        Gdx.graphics.getWidth() / 2f - level1Button.getWidth() / 2 - level1Button.getWidth() * 1.5f,
-        Gdx.graphics.getHeight() / 2f - level1Button.getHeight() / 2);
-    level2Button = new TextButton("Level 2", buttonStyle);
-    level2Button.setPosition(
-        Gdx.graphics.getWidth() / 2f - level1Button.getWidth() / 2,
-        Gdx.graphics.getHeight() / 2f - level1Button.getHeight() / 2);
-    level3Button = new TextButton("Level 3", buttonStyle);
-    level3Button.setPosition(
-        Gdx.graphics.getWidth() / 2f - level1Button.getWidth() / 2 + level3Button.getWidth() * 1.5f,
-        Gdx.graphics.getHeight() / 2f - level1Button.getHeight() / 2);
-    stage.addActor(level1Button);
-    stage.addActor(level2Button);
-    stage.addActor(level3Button);
-    level1Button.addListener(
+    buttonStyle.up = skin.getDrawable("black_alpha_mid");
+    buttonStyle.down = skin.getDrawable("black_alpha_mid");
+    buttonStyle.checked = skin.getDrawable("black_alpha_mid");
+
+    creditsButton = new TextButton("Credits", buttonStyle);
+    creditsButton.setPosition(
+        Gdx.graphics.getWidth() / 2f - creditsButton.getWidth() / 2,
+        Gdx.graphics.getHeight() / 3f + creditsButton.getHeight() / 2);
+    creditsButton.addListener(
         new ChangeListener() {
           @Override
           public void changed(ChangeEvent event, Actor actor) {
-            // game.setScreen(new GameScreen(game, 1));
-            game.setScreen(new TutorialScreen(game, "game1"));
+            game.setScreen(new CreditsScreen(game));
           }
         });
-    level1Button.addListener(
+    stage.addActor(creditsButton);
+
+    titleButton = new TextButton("Title", buttonStyle);
+    titleButton.setPosition(
+        Gdx.graphics.getWidth() / 2f - titleButton.getWidth() / 2,
+        Gdx.graphics.getHeight() / 3f - titleButton.getHeight() / 2);
+    titleButton.addListener(
         new ChangeListener() {
           @Override
           public void changed(ChangeEvent event, Actor actor) {
-            // game.setScreen(new GameScreen(game, 2));
+            game.setScreen(new TitleScreen(game));
           }
         });
-    level1Button.addListener(
-        new ChangeListener() {
-          @Override
-          public void changed(ChangeEvent event, Actor actor) {
-            // game.setScreen(new GameScreen(game, 3));
-          }
-        });
+    stage.addActor(titleButton);
   }
 
   @Override
@@ -116,45 +104,26 @@ public class LevelSelectorScreen extends ScreenAdapter {
             bg.getHeight() * bgScaleFactor);
     font.draw(
         game.getBatch(),
-        "LEVEL SELECTION",
+        "CONGRATULATIONS!\nYou completed the game in " + completionTime + " seconds!",
         winWidth / 2f - winWidth / 10f,
         winHeight / 2f + winHeight / 5f,
         winWidth / 5f,
         1,
         false);
-    stage.draw();
-    game.getBatch()
-        .draw(
-            lock,
-            level2Button.getX(),
-            level2Button.getY(),
-            level2Button.getWidth(),
-            level2Button.getHeight());
-    game.getBatch()
-        .draw(
-            lock,
-            level3Button.getX(),
-            level3Button.getY(),
-            level3Button.getWidth(),
-            level3Button.getHeight());
     game.getBatch().end();
+    stage.draw();
   }
 
   @Override
   public void resize(int width, int height) {
     super.resize(width, height);
-    level1Button.setPosition(
-        width / 2f - level1Button.getWidth() / 2 - level1Button.getWidth() * 1.5f,
-        height / 2f - level1Button.getHeight() / 2);
-    level2Button.setPosition(
-        width / 2f - level1Button.getWidth() / 2, height / 2f - level1Button.getHeight() / 2);
-    level3Button.setPosition(
-        width / 2f - level1Button.getWidth() / 2 + level3Button.getWidth() * 1.5f,
-        height / 2f - level1Button.getHeight() / 2);
+    creditsButton.setPosition(
+        width / 2f - creditsButton.getWidth() / 2, height / 3f + creditsButton.getHeight() / 2);
+    titleButton.setPosition(
+        width / 2f - titleButton.getWidth() / 2, height / 3f - titleButton.getHeight() / 2);
     stage.clear();
-    stage.addActor(level1Button);
-    stage.addActor(level2Button);
-    stage.addActor(level3Button);
+    stage.addActor(creditsButton);
+    stage.addActor(titleButton);
     stage.getViewport().update(width, height);
     camera.setToOrtho(false, width, height);
   }
@@ -166,7 +135,6 @@ public class LevelSelectorScreen extends ScreenAdapter {
     batch.dispose();
     font.dispose();
     bg.dispose();
-    lock.dispose();
     stage.dispose();
     skin.dispose();
     atlas.dispose();
