+++ desktop/src/com/neves6/piazzapanic/DesktopLauncher.java
@@ -1,7 +1,7 @@
 package com.neves6.piazzapanic;
 
 import com.badlogic.gdx.backends.lwjgl3.Lwjgl3Application;
-import com.badlogic.gdx.backends.lwjgl3.Lwjgl3ApplicationConfiguration;
+import com.badlogic.gdx.backends.lwjgl3.Lwjgl3ApplicationConfiguration;import com.neves6.piazzapanic.screens.PiazzaPanicGame;
 
 // Please note that on macOS your application needs to be started with the -XstartOnFirstThread JVM
 // argument
